requirements.txt
fastapi==0.107.0
uvicorn==0.26.0
gunicorn==21.2.0
# Removed python-whois as we're using socket-based WHOIS queries
dnspython==2.6.0
requests==2.31.0
python-dotenv==1.0.0
python-dateutil==2.8.2
pydantic==1.10.13
